add Handler

+++ src/main/java/cn/coder/struts/StrutsFilter.java
@@ -1,11 +1,11 @@
 package cn.coder.struts;
 
 import java.io.IOException;
+import java.util.ArrayList;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
-import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
@@ -15,9 +15,10 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import cn.coder.struts.core.Action;
-import cn.coder.struts.core.ActionHandler;
+import cn.coder.struts.core.Handlers;
 import cn.coder.struts.core.StrutsContext;
+import cn.coder.struts.view.Action;
+import cn.coder.struts.wrapper.ActionWrapper;
 
 /**
  * 核心控制类<br>
@@ -29,16 +30,18 @@ import cn.coder.struts.core.StrutsContext;
  */
 public class StrutsFilter implements Filter {
 	private static final Logger logger = LoggerFactory.getLogger(StrutsFilter.class);
-	private StrutsContext context = new StrutsContext();
-	private ActionHandler handler;
+	private StrutsContext context;
+	private ActionWrapper actionWrapper;
+	private ArrayList<Class<?>> handlerStack;
 
 	@Override
 	public void init(FilterConfig filterConfig) throws ServletException {
 		long start = System.currentTimeMillis();
-		ServletContext servletContext = filterConfig.getServletContext();
-		context.init(servletContext);
-		context.startUp(servletContext);
-		handler = context.getHandler();
+		context = new StrutsContext();
+		context.init(filterConfig.getServletContext());
+		context.start();
+		actionWrapper = context.getWrapper();
+		handlerStack = context.getHandlers();
 		if (logger.isDebugEnabled())
 			logger.debug("Struts context started with {} ms", (System.currentTimeMillis() - start));
 	}
@@ -48,32 +51,33 @@ public class StrutsFilter implements Filter {
 			throws IOException, ServletException {
 		HttpServletRequest req = (HttpServletRequest) request;
 		HttpServletResponse res = (HttpServletResponse) response;
-		req.setCharacterEncoding("UTF-8");
-		res.setCharacterEncoding("UTF-8");
+		req.setCharacterEncoding("utf-8");
 
-		if ("OPTIONS".equals(req.getMethod())) {
-			chain.doFilter(request, response);
+		Handlers handlers = new Handlers(handlerStack, req, res);
+		if (!handlers.complete()) {
+			if (logger.isDebugEnabled())
+				logger.debug("Handler stoped");
 			return;
 		}
 
-		String servletPath = req.getServletPath();
-		Action action = handler.getAction(servletPath);
+		Action action = actionWrapper.getAction(req);
 		if (action != null) {
-			if (logger.isDebugEnabled())
-				logger.debug("Find the path '{}'", servletPath);
-			handler.handle(action, req, res);
+			actionWrapper.execute(action, req, res);
 		} else {
-			if (logger.isDebugEnabled())
-				logger.debug("Not found the path '{}'", servletPath);
 			chain.doFilter(request, response);
 		}
 	}
 
 	@Override
 	public void destroy() {
-		context.destroy();
-		if (logger.isDebugEnabled())
-			logger.debug("Struts context destroied");
+		if (handlerStack != null) {
+			handlerStack.clear();
+			handlerStack = null;
+		}
+		if (context != null) {
+			context.destroy();
+			context = null;
+		}
 	}
 
 }
